Move code for storeNamesAndProperties
